feat(contract): dcarbon-1207558596855211 fix minting

+++ programs/dcarbon-contract/src/instructions/config/set_minting_fee.rs
@@ -2,7 +2,7 @@ use anchor_lang::prelude::*;
 
 use crate::instructions::SetConfig;
 
-pub fn set_minting_fee(ctx: Context<SetConfig>, minting_fee: u64) -> Result<()> {
+    pub fn set_minting_fee(ctx: Context<SetConfig>, minting_fee: f64) -> Result<()> {
     let contract_config = &mut ctx.accounts.contract_config;
 
     contract_config.minting_fee = minting_fee;
